+++ forte2/dsrg/__init__.py
@@ -1,2 +1,2 @@
+from .rel_dsrg_mrpt2 import RelDSRG_MRPT2
 from .dsrg_mrpt2 import DSRG_MRPT2
-from .sf_dsrg_mrpt2 import SF_DSRG_MRPT2
